move embeddable content into smv in ml to remove duplication
